feat: CLIN-294 -  clean-up + _tag support
